% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{cluster_embedding}
\alias{cluster_embedding}
\title{Cluster the resulting embedding}
\usage{
cluster_embedding(sce, k = NULL, method = c("kmeans", "mm"))
}
\arguments{
\item{sce}{The SCESet object}

\item{k}{The number of clusters to find in the data}

\item{method}{Either 'kmeans' or 'mm' to use \code{mclust}}
}
\value{
The dataframe M with a new numeric variable `cluster` containing the assigned cluster
}
\description{
Cluster the embedded representation using either kmeans or mixture models
from the mclust package
}

