% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{compare_models}
\alias{compare_models}
\title{Perform likelihood ratio test}
\usage{
compare_models(model, null_model)
}
\arguments{
\item{model}{The full model y ~ pseudotime}

\item{null_model}{The null model y ~ 1}
}
\value{
The p-value
}
\description{
Perform likelihood ratio test
}

