% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr-package.R
\docType{package}
\name{embeddr-package}
\alias{embeddr-package}
\title{Pseudotemporal ordering of single-cell RNA-seq data using laplacian eigenmaps and principal curves}
\description{
\pkg{embeddr} provides a set of methods for manifold learning on single-cell data followed by
principal curve fitting to assign a pseudotime to cells.
}

