% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{fit_null_model}
\alias{fit_null_model}
\title{Fit the null pseudotime model}
\usage{
fit_null_model(sce, gene)
}
\arguments{
\item{sce}{The SCESet object}

\item{gene}{The gene name to fit}
}
\value{
An object of class VGAM
}
\description{
This function fits the null expression profile in y (ie y ~ 1). A tobit model is
used to censor values less than min_expr
}

