% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{fit_pseudotime_model}
\alias{fit_pseudotime_model}
\title{Fit the gene expression profile in pseudotime}
\usage{
fit_pseudotime_model(sce, gene, ...)
}
\arguments{
\item{sce}{An object of type SCESet}

\item{gene}{The gene name to fit, or anything that can subset the SCE model to a single gene}
}
\value{
An object of class AER::tobit
}
\description{
This function fits the expression profile of y as a function of
pseudotime using a natural cubic spline of degree three. A tobit model
is used to censor values less than min_expr
}
\examples{
library(scater)
data('sc_example_counts') ; sce <- newSCESet(countData = sc_example_counts)
sce <- embeddr(sce)
sce <- fit_pseudotime(sce)
model <- fit_pseudotime_model(sce, 1) # fit for first gene
}

