% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{fit_pseudotime_models}
\alias{fit_pseudotime_models}
\title{Generate a list of pseudotime models corresponding to ALL genes in sce}
\usage{
fit_pseudotime_models(sce, n_cores = 2)
}
\description{
\strong{WARNING}: The list returned can be huge, so use only on a few genes.
}

