% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{gene_pseudotime_test}
\alias{gene_pseudotime_test}
\title{Test a single gene as a function of pseudotime}
\usage{
gene_pseudotime_test(sce, gene, full_model = NULL)
}
\arguments{
\item{sce}{An object of class SCESet}

\item{gene}{Name or index of the gene to be tested.}

\item{full_model}{If a full pseudotime model has already been calculated, use that. Otherwise (if NULL, default),
recalculate one on the fly.}
}
\value{
A p-value for the given gene
}
\description{
Test a single gene as a function of pseudotime
}
\examples{
library(scater)
data('sc_example_counts') ; sce <- newSCESet(countData = sc_example_counts)
sce <- embeddr(sce)
sce <- fit_pseudotime(sce)
pval <- gene_pseudotime_test(sce, 1) # select first gene
}

