% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{laplacian_eigenmap}
\alias{laplacian_eigenmap}
\title{Laplacian eigenmaps}
\usage{
laplacian_eigenmap(W, measure_type = c("unorm", "norm"), p = 2)
}
\arguments{
\item{W}{The weighted graph adjacency matrix}

\item{measure_type}{Type of laplacian eigenmap (norm for normalised, unorm otherwise)}

\item{p}{Dimension of the embedded space, default is 2}
}
\value{
The p-dimensional embedding
}
\description{
Construct a laplacian eigenmap embedding
}

