% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{log_expression}
\alias{log_expression}
\title{Log the expression of a given SCESet}
\usage{
log_expression(sce, base = 10, pseudocount = NULL)
}
\arguments{
\item{sce}{An object of class SCESet}

\item{base}{The base of the logarithm}

\item{pseudocount}{The count to add to the expression to avoid taking
the logarithms of negative numbers. If NULL this will
be worked out automatically.}
}
\value{
An SCESet with the expression logged
}
\description{
Log the expression of a given SCESet
}

