% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot the nearest neighbour graph in the embedding}
\usage{
plot_graph(sce)
}
\value{
A ggplot graphic
}
\description{
Plot the nearest neighbour graph in the embedding
}

