% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{plot_pseudotime_density}
\alias{plot_pseudotime_density}
\title{Plot density of cells in pseudotime}
\usage{
plot_pseudotime_density(sce)
}
\arguments{
\item{sce}{An object of class SCESet}
}
\value{
A `ggplot` object
}
\description{
Plot density of cells in pseudotime
}

