% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{plot_pseudotime_metrics}
\alias{plot_pseudotime_metrics}
\title{Plot metrics in pseudotime}
\usage{
plot_pseudotime_metrics(sce, gene, window_size = NULL)
}
\value{
A ggplot graphic
}
\description{
Plot various metrics (mean, variance, CV2 and signal-to-noise ratio) using
a sliding window approach
}

