% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{predicted_expression}
\alias{predicted_expression}
\title{Create a predicted expression matrix}
\usage{
predicted_expression(sce, models = NULL, n_cores = 2)
}
\description{
Create a predicted expression matrix
}

