% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{remove_pseudotime}
\alias{remove_pseudotime}
\title{removes the pseudotime and low dimensional trajectory}
\usage{
remove_pseudotime(sce)
}
\arguments{
\item{sce}{The \code{SCESet}}
}
\value{
An \code{SCESet} with the pseudotime & trajectory removed
}
\description{
removes the pseudotime and low dimensional trajectory
}

