% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{reverse_pseudotime}
\alias{reverse_pseudotime}
\title{Reverse pseudotime}
\usage{
reverse_pseudotime(sce)
}
\arguments{
\item{sce}{An object of class \code{SCESet}}
}
\value{
\code{sce} with the pseudotime reversed.
}
\description{
Reverse the pseudotimes of cells via pseudotime x -> -x + max(x) + min(x)
}

