% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{weighted_graph}
\alias{weighted_graph}
\title{Construct a weighted graph adjacency matrix}
\usage{
weighted_graph(x, kernel = c("nn", "dist", "heat"),
  metric = c("correlation", "euclidean", "cosine"),
  nn = round(log(ncol(x))), eps = NULL, t = NULL, symmetrize = c("mean",
  "ceil", "floor"))
}
\arguments{
\item{x}{The feature-by-sample (e.g. genes are rows, cells are columns) data matrix}

\item{kernel}{The choice of kernel. 'nn' will give nearest neighbours, 'dist' gives minimum distance and
'heat' gives a heat kernel. Discussed in detail in 'Laplacian Eigenmaps and Spectral Techniques for Embedding and Clustering',
Belkin & Niyogi}

\item{metric}{The metric with which to assess 'closeness' for nearest neighbour selection, one of
'correlation' (pearson) or 'euclidean'. Default is 'correlation'.}

\item{nn}{Number of nearest neighbours if kernel == 'nn'}

\item{eps}{Maximum distance parameter if kernel == 'dist'}

\item{t}{'time' for heat kernel if kernel == 'heat'}

\item{symmetrize}{How to make the adjacency matrix symmetric. Note that slightly
counterintuitively, node i having node j as a nearest neighbour doesn't guarantee node
j has node i. There are several ways to get round this:
\itemize{
\item \code{mean} If the above case occurs make the link weight 0.5 so the adjacency matrix becomes \eqn{0.5(A + A')}
\item \code{ceil} If the above case occurs set the link weight to 1 (ie take the ceiling of the mean case)
\item \code{floor} If the above case occurs set the link weight to 0 (ie take the floor of the mean case)
}}
}
\value{
An n by n adjacency matrix
}
\description{
Construct a weighted graph adjacency matrix
}
\examples{
x <- matrix(rnorm(50), ncol = 10) # synthetic data matrix
wg <- weighted_graph(x)
}

