% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate metrics through pseudotime}
\usage{
calculate_metrics(sce, gene, window_size = NULL)
}
\arguments{
\item{sce}{A \code{SCESet} object}

\item{gene}{The gene of interest on which to calculate the metrics}

\item{window_size}{The size of the sliding window. By default taken to be half the number of cells}
}
\value{
An object of class `data.frame` where each column is a metric (window-averaged pseudotime,
mean, variance, CV2, signal to noise ratio)
}
\description{
Calculate various metrics (mean, variance, CV2 and signal-to-noise ratio) using
a sliding window approach
}

