% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{cluster_embedding}
\alias{cluster_embedding}
\title{Cluster the resulting embedding}
\usage{
cluster_embedding(sce, method = c("mm", "kmeans"), k = NULL)
}
\arguments{
\item{sce}{The SCESet object}

\item{method}{Either 'kmeans' or 'mm' to use \code{mclust}}

\item{k}{The number of clusters to find in the data}
}
\value{
The dataframe M with a new numeric variable `cluster` containing the assigned cluster
}
\description{
Cluster the embedded representation using either kmeans or mixture models
from the mclust package
}
\examples{
library(scater)
data('sc_example_counts') ; sce <- newSCESet(countData = sc_example_counts)
sce <- embeddr(sce)
sce <- cluster_embedding(sce)
}

