% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{fit_pseudotime}
\alias{fit_pseudotime}
\title{Fit the pseudotime curve}
\usage{
fit_pseudotime(sce, clusters = NULL, ...)
}
\arguments{
\item{sce}{The SCESet object}

\item{clusters}{The (numeric) clusters to use for the curve fitting. If NULL (default) then
all points are used}

\item{...}{Additional arguments to be passed to \code{princurve} from \pkg{principal.curve}.}
}
\value{
The dataframe \code{M} with three new variables:
\describe{
\item{pseudotime}{The pseudotime of the cell (arc-length from beginning of curve)}
\item{trajectory_1}{The x-coordinate of a given cell's projection onto the curve}
\item{trajectory_2}{The y-coordinate of a given cell's projection onto the curve}}
}
\description{
Fits the pseudotime curve using principal curves from the princurve library
}

