% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{fit_pseudotime_models}
\alias{fit_pseudotime_models}
\title{Generate a list of pseudotime models corresponding to ALL genes in sce}
\usage{
fit_pseudotime_models(sce, n_cores = 2)
}
\arguments{
\item{sce}{An object of class \code{SCESet}}

\item{n_cores}{The number of cores to use in the call to \code{mclapply}}
}
\description{
This will iterate over every gene in the \code{SCESet} and produce a model fit
for each. \strong{WARNING}: The list returned can be huge, so use only on a few genes.
}

