% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{gene_pseudotime_test}
\alias{gene_pseudotime_test}
\title{Test a single gene as a function of pseudotime}
\usage{
gene_pseudotime_test(gene_name, sce, full_model = NULL)
}
\arguments{
\item{gene_name}{Name of the gene to be tested. Must be consistent with `featureNames(sce)`}

\item{sce}{An object of class SCESet}

\item{full_model}{If a full pseudotime model has already been calculated, use that. Otherwise (if NULL, default),
recalculate one on the fly.}
}
\description{
Test a single gene as a function of pseudotime
}

