% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{plot_degree_dist}
\alias{plot_degree_dist}
\title{Plot the degree distribution of the weight matrix}
\usage{
plot_degree_dist(W, ignore_weights = FALSE)
}
\arguments{
\item{ignore_weights}{If TRUE weights are discretised to 0 - 1}

\item{The}{SCESet object}
}
\value{
A ggplot histogram of weights
}
\description{
Plot the degree distribution of the weight matrix
}

