% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot the cells in the embedding}
\usage{
plot_embedding(sce, color_by = "cluster")
}
\arguments{
\item{sce}{The SCESet object}

\item{color_by}{The variable to color the embedding with (defaults to cluster)}
}
\value{
A \code{ggplot2} plot
}
\description{
This function takes a data frame with at least positional (component_0 & component_1) information
and plots the resulting embedding. If clusters are assigned it can colour by these, and if a pseudotime
trajectory is assigned it will plot this through the embeddding.
}

