% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot the nearest neighbour graph in the embedding}
\usage{
plot_graph(sce)
}
\arguments{
\item{sce}{An object of class \code{SCESet}}
}
\value{
A ggplot graphic
}
\description{
This plots the cells in the reduced dimension and connects cells if they are
connected in the nearest neighbour graph W. A vertice between cells i and j is
coloured red if both i and j are nearest neigbours of each other and grey otherwise.
}

