% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{plot_in_pseudotime}
\alias{plot_in_pseudotime}
\title{Plot cells in pseudotime}
\usage{
plot_in_pseudotime(sce, nrow = NULL, ncol = NULL, use_short_names = TRUE)
}
\arguments{
\item{nrow}{Number of rows of plots; passed to \code{facet_wrap}}

\item{ncol}{Number of columns of plots; passed to \code{facet_wrap}}

\item{Mp}{A dataframe containing the embedding}

\item{xp}{The gene-by-cell matrix of log normalised expression counts}

\item{genes}{The genes to use for the embedding}

\item{short_names}{Short gene names to display; default NULL and \code{genes} is used for gene names}
}
\value{
A \code{ggplot2} plot
}
\description{
Plot a set of genes through pseudotime
}

