% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{plot_pseudotime_density}
\alias{plot_pseudotime_density}
\title{Plot density of cells in pseudotime}
\usage{
plot_pseudotime_density(sce, color_by = NULL, reverse = FALSE)
}
\arguments{
\item{sce}{An object of class SCESet}

\item{color_by}{The variable (in \code{pData(sce)}) by which to colour the density plot.}

\item{reverse}{Logical If true the pseudotime will be reversed.}
}
\value{
A `ggplot` object
}
\description{
This returns a \code{ggplot} density plot of the cells across pseudotime.
}
\examples{
library(scater)
data('sc_example_counts') ; sce <- newSCESet(countData = sc_example_counts)
sce <- embeddr(sce)
sce <- fit_pseudotime(sce)
plot_pseudotime_density(sce)
}

