% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{plot_pseudotime_metrics}
\alias{plot_pseudotime_metrics}
\title{Plot metrics in pseudotime}
\usage{
plot_pseudotime_metrics(sce, ...)
}
\arguments{
\item{sce}{An object of class \code{SCESet}}

\item{...}{Additional arguments to be passed to \code{calculate_metrics} concerning
how the metrics are calculated.}
}
\value{
A ggplot graphic
}
\description{
Plot various metrics (mean, variance, CV2 and signal-to-noise ratio) using
a sliding window approach. Calculates metrics if not already done so.
}
\examples{
library(scater)
data('sc_example_counts') ; sce <- newSCESet(countData = sc_example_counts)
sce <- embeddr(sce)
sce <- fit_pseudotime(sce)
plot_pseudotime_metrics(sce)
}

