% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{predicted_expression}
\alias{predicted_expression}
\title{Create a predicted expression matrix}
\usage{
predicted_expression(sce, models = NULL, n_cores = 2)
}
\arguments{
\item{sce}{An object of class \code{SCESet}}

\item{models}{An object representing models. If of type \code{list} then for each element
the predicted expression is computed and a matrix returned. If of type \code{model} for
which a \code{predict} function is available, then a single vector corresponding to
\code{predict(model)} is returned. If NULL then the model is computed for all genes in
\code{sce} and the resulting list returned.}

\item{n_cores}{The number of cores to pass to \code{mclapply}.}
}
\value{
A dataframe of predicted expression where rows are cells (pseudotime) and
columns are genes
}
\description{
Given a list of models return a matrix corresponding to the prediction from the models.
Each column represents a gene and each row its expression at a given point in pseudotime.
}
\examples{
library(scater)
data('sc_example_counts') ; sce <- newSCESet(countData = sc_example_counts)
sce <- embeddr(sce)
sce <- fit_pseudotime(sce)
pe <- predicted_expression(sce[1:4,]) # use first four genes
}

