% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{pseudotime}
\alias{pseudotime}
\title{Retrieve the pseudotime assignment from sce}
\usage{
pseudotime(sce)
}
\value{
A numeric vector of pseudotimes
}
\description{
Retrieve the pseudotime assignment from sce
}

