% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/embeddr.R
\name{reverse_pseudotime}
\alias{reverse_pseudotime}
\title{Reverse pseudotime}
\usage{
reverse_pseudotime(sce)
}
\arguments{
\item{M}{A dataframe containing a pseudotime variable to be reversed}
}
\value{
A dataframe with the reversed pseudotime
}
\description{
Reverse the pseudotimes of cells
}

